instances: reject reserved build volume IDs in attachments
The reserved internal volume ID prefixes (build-cache-, build-disk-,
build-config-, build-source-) were already rejected by the volume
create/delete APIs, but CreateInstance accepted arbitrary volume ID
attachments. A caller could attach another scope's persistent BuildKit
cache volume read-write at a non-system mount path, tampering with or
holding that scope's builds. validateVolumeAttachments now rejects
reserved IDs unless the instance is internal
(AllowSystemVolumeMounts, which is never settable from the API).

The prefix list moves from lib/builds to lib/volumes so lib/instances
can consult it without an import cycle (lib/builds imports
lib/instances).
